Order_Processor/requirements.txt
httpx
pydantic
structlog
PyYAML
pytz
loguru
python-dotenv
pandas
watchdog